+++ crates/cli/lib/utils.ts
@@ -1,9 +1,6 @@
-import { createRequire } from 'node:module';
-import { Binary } from './binary-install';
 import os from 'node:os';
-import path from 'node:path';
 
-function getPlatform() {
+export function getPlatform() {
   const type = os.type();
   const arch = os.arch();
 
@@ -43,19 +40,4 @@ function getPlatform() {
   }
 
   throw new Error(`Unsupported platform: ${type} ${arch}`);
-}
-
-export function getBinary() {
-  const { platform, name } = getPlatform();
-  const customRequire = createRequire(import.meta.url);
-
-  const { name: packageName, version } = customRequire('../package.json') as {
-    name: string;
-    version: string;
-  };
-
-  const url = `https://github.com/lagossapp/lagoss/releases/download/${packageName}@${version}/${platform}.tar.gz`;
-  const installationDirectory = path.join(os.homedir(), '.lagoss');
-
-  return new Binary(url, { name, installationDirectory, version });
 }