+++ crates/cli/src/commands/logout.rs
@@ -2,9 +2,7 @@ use crate::utils::{get_theme, Config};
 use anyhow::{anyhow, Result};
 use dialoguer::{console::style, Confirm};
 
-pub fn logout() -> Result<()> {
-    let mut config = Config::new()?;
-
+pub fn logout(config: &Config) -> Result<()> {
     if config.token.is_none() {
         return Err(anyhow!("You are not logged in."));
     }
@@ -15,6 +13,7 @@ pub fn logout() -> Result<()> {
         .interact()?
     {
         true => {
+            let mut config = config.clone();
             config.set_token(None);
             config.save()?;