+++ packages/dashboard/server/lib/redis.ts
@@ -1,8 +1,39 @@
 import { createClient, RedisClientType } from 'redis';
 
-let redis: RedisClientType;
+let redis: RedisClient;
 
-export async function useRedis(): Promise<RedisClientType> {
+type DeployPayload = {
+  functionId: string;
+  functionName: string;
+  deploymentId: string;
+  domains: string[];
+  memory: number;
+  tickTimeout: number;
+  totalTimeout: number;
+  cron: string | null;
+  cronRegion: string | null;
+  env: Record<string, string>;
+  isProduction: boolean;
+  assets: string[];
+};
+
+type UndeployPayload = DeployPayload;
+
+type PromotePayload = DeployPayload & {
+  previousDeploymentId: string;
+};
+
+type Payloads = {
+  deploy: DeployPayload;
+  undeploy: UndeployPayload;
+  promote: PromotePayload;
+};
+
+type RedisClient = Omit<RedisClientType, 'publish'> & {
+  publish: <K extends keyof Payloads>(channel: K, message: Payloads[K]) => Promise<number>;
+};
+
+export async function useRedis(): Promise<RedisClient> {
   if (redis) {
     return redis;
   }
@@ -16,7 +47,14 @@ export async function useRedis(): Promise<RedisClientType> {
 
   await redisClient.connect();
 
-  redis = redisClient as RedisClientType;
+  Object.assign(redisClient, {
+    publish: async (channel: string, payload: unknown) => {
+      const message = typeof payload === 'string' ? payload : JSON.stringify(payload);
+      return redisClient.publish(channel, message);
+    },
+  });
+
+  redis = redisClient as RedisClient;
 
   return redis;
 }