+++ crates/cli/CHANGELOG.md
@@ -1,5 +1,60 @@
 # @lagoss/cli
 
+## [0.13.0](https://github.com/lagossapp/lagoss/compare/@lagoss/cli@0.12.3...@lagoss/cli@0.13.0) (2025-10-31)
+
+
+### Features
+
+* support passing app id or name to cli ([#237](https://github.com/lagossapp/lagoss/issues/237)) ([ed84c14](https://github.com/lagossapp/lagoss/commit/ed84c14f532a49f7f2ea5de0efa0376ba35192b2))
+
+## [0.12.3](https://github.com/lagossapp/lagoss/compare/@lagoss/cli@0.12.2...@lagoss/cli@0.12.3) (2025-10-30)
+
+
+### Bug Fixes
+
+* cli release and improve js wrapper ([#235](https://github.com/lagossapp/lagoss/issues/235)) ([592e79a](https://github.com/lagossapp/lagoss/commit/592e79a80e5b467ca8ff04fb64ddc84b913c898f))
+
+## [0.12.2](https://github.com/lagossapp/lagoss/compare/@lagoss/cli@0.12.1...@lagoss/cli@0.12.2) (2025-10-30)
+
+
+### Bug Fixes
+
+* add it-token for cli   ([#233](https://github.com/lagossapp/lagoss/issues/233)) ([74d6996](https://github.com/lagossapp/lagoss/commit/74d699631388eb7881b505fecd9a58c1caed872a))
+
+## [0.12.1](https://github.com/lagossapp/lagoss/compare/@lagoss/cli@0.12.0...@lagoss/cli@0.12.1) (2025-10-30)
+
+
+### Bug Fixes
+
+* disable git checks for pnpm publish ([#229](https://github.com/lagossapp/lagoss/issues/229)) ([2c97411](https://github.com/lagossapp/lagoss/commit/2c974111d224dbe85a7b56ca7b40aaca6828542e))
+
+## [0.12.0](https://github.com/lagossapp/lagoss/compare/@lagoss/cli@0.11.0...@lagoss/cli@0.12.0) (2025-10-30)
+
+
+### Features
+
+* add env vars for cli config and improve arguments + output ([#211](https://github.com/lagossapp/lagoss/issues/211)) ([1ea69b5](https://github.com/lagossapp/lagoss/commit/1ea69b5f34acae34ec53c004eac7f9bd29eff53d))
+* support static dir uploading and detect app folder ([#213](https://github.com/lagossapp/lagoss/issues/213)) ([e14021d](https://github.com/lagossapp/lagoss/commit/e14021ddaa15cf8d5df3b2eab5fd2a55fd01ebcf))
+
+
+### Bug Fixes
+
+* **deps:** update rust crate console to 0.16.0 ([#185](https://github.com/lagossapp/lagoss/issues/185)) ([25ca7aa](https://github.com/lagossapp/lagoss/commit/25ca7aaee7e17661b289919ce6f375f0a9f86d40))
+* **deps:** update rust crate dialoguer to 0.12.0 ([#186](https://github.com/lagossapp/lagoss/issues/186)) ([06e89d5](https://github.com/lagossapp/lagoss/commit/06e89d521ab3eaeeee302179720f529eab084abe))
+
+## [0.11.0](https://github.com/lagossapp/lagoss/compare/@lagoss/cli@0.10.2...@lagoss/cli@0.11.0) (2025-10-08)
+
+### Features
+
+- use rest api with cli ([#161](https://github.com/lagossapp/lagoss/issues/161)) ([de3b00e](https://github.com/lagossapp/lagoss/commit/de3b00ebba75b4cba7dcfcdd8030d535c83eb447))
+- use rest api with cli ([#193](https://github.com/lagossapp/lagoss/issues/193)) ([4581e10](https://github.com/lagossapp/lagoss/commit/4581e103afe447c9006c71cf43a332da00b4db92))
+
+### Bug Fixes
+
+- correctly bundle cli ([#203](https://github.com/lagossapp/lagoss/issues/203)) ([28c8a7b](https://github.com/lagossapp/lagoss/commit/28c8a7b549fab51ed6083177d200873423d22c5f))
+- **deps:** update dependency esbuild to ^0.25.0 [security] ([#144](https://github.com/lagossapp/lagoss/issues/144)) ([1e5ed21](https://github.com/lagossapp/lagoss/commit/1e5ed21b83f83f84067878cab86bac6ea999772c))
+- **deps:** update rust crate flume to 0.11.0 ([#188](https://github.com/lagossapp/lagoss/issues/188)) ([4ee9ddb](https://github.com/lagossapp/lagoss/commit/4ee9ddb02b95ed7c1abc986f3cafa1ee25fc934b))
+
 ## 0.10.2
 
 ### Patch Changes